Добавлена возможность сброса стейтов

+++ tg/handlers.py
@@ -1,21 +1,26 @@
-from settings import logger
 import json
 from datetime import datetime
+
+import settings
+from settings import logger
 
 from aiogram import Dispatcher, types
 from aiogram.dispatcher import FSMContext
 from aiogram.dispatcher.filters import Text
 
 from tg import buttons
+from tg.decorators import ascii_only_message
 from tg.states import UserSingleState, UserPeriodicState, CreateUserState, DeleteUserState
 from scheduler import scheduler
 from parser.start_parse import start_parse
-from parser.login.base_session import BaseSession
 
 NOTIFIED_TRANSACTION = []
 
 
 def register_handlers_requests(dispatcher: Dispatcher):
+    dispatcher.register_message_handler(cancel_handler, commands=["cancel"], state="*")
+
+    dispatcher.register_message_handler(cancel_handler, Text(equals='Отмена', ignore_case=True), state="*")
 
     dispatcher.register_message_handler(get_name_for_single, Text(equals='Получить список активных заявок',
                                                                   ignore_case=True))
@@ -27,12 +32,12 @@ def register_handlers_requests(dispatcher: Dispatcher):
 
     dispatcher.register_message_handler(start_periodic_parse, state=UserPeriodicState.user_name)
 
-    dispatcher.register_message_handler(cancel_schedule_task, Text(equals='Остановить',
-                                                                   ignore_case=True))
+    dispatcher.register_message_handler(cancel_periodic_parse, Text(equals='Остановить',
+                                                                    ignore_case=True))
 
-    dispatcher.register_message_handler(help_response, commands=["help"],)
+    dispatcher.register_message_handler(help_response, commands=["help"])
 
-    dispatcher.register_message_handler(start_response, commands=["start"],)
+    dispatcher.register_message_handler(start_response, commands=["start"])
 
     dispatcher.register_message_handler(start_user_registration, commands=["adduser"], state=None)
 
@@ -42,9 +47,11 @@ def register_handlers_requests(dispatcher: Dispatcher):
 
     dispatcher.register_message_handler(load_name, state=CreateUserState.name)
 
-    dispatcher.register_message_handler(start_delete_user, commands=["deleteuser"], state=None)
+    dispatcher.register_message_handler(delete_user_start, commands=["deleteuser"], state=None)
 
-    dispatcher.register_message_handler(delete_user, state=DeleteUserState.name)
+    dispatcher.register_message_handler(delete_user_confirm, state=DeleteUserState.user_name)
+
+    dispatcher.register_message_handler(delete_user_finish, state=DeleteUserState.confirm)
 
 
 async def get_name_for_periodic(message: types.Message):
@@ -76,9 +83,13 @@ async def start_periodic_parse(message: types.Message, state: FSMContext):
 async def start_single_parse(message: types.Message, state: FSMContext):
     await state.update_data(user_name=message.text)
     data = await state.get_data()
-    await message.reply(text=f"Запуск одиночной проверки. Пользователь: {data.get('user_name')}", reply_markup=buttons.main_menu)
-    await get_requests(message, user_name=data.get('user_name'))
-    await state.finish()
+    try:
+        await message.reply(text=f"Запуск одиночной проверки. Пользователь: {data.get('user_name')}", reply_markup=buttons.main_menu)
+        await get_requests(message, user_name=data.get('user_name'))
+    except json.decoder.JSONDecodeError:
+        await message.answer("Пользователь не существует")
+    finally:
+        await state.finish()
 
 
 async def get_requests(message: types.Message, user_name: str, is_schedule: bool = False):
@@ -142,72 +153,83 @@ async def get_requests(message: types.Message, user_name: str, is_schedule: bool
 
 async def start_user_registration(message: types.Message):
     await CreateUserState.login.set()
-    await message.reply('Введите логин')
+    await message.reply('Введите логин', reply_markup=buttons.cancel_menu)
 
 
+@ascii_only_message
 async def load_login(message: types.Message, state: FSMContext):
     await state.update_data(login=message.text)
     await CreateUserState.password.set()
-    await message.reply('Введите пароль')
+    await message.reply('Введите пароль', reply_markup=buttons.cancel_menu)
 
 
+@ascii_only_message
 async def load_password(message: types.Message, state: FSMContext):
     await state.update_data(password=message.text)
     await CreateUserState.name.set()
-    await message.reply('Введите отображаемое имя')
+    await message.reply('Введите отображаемое имя', reply_markup=buttons.cancel_menu)
 
 
+@ascii_only_message
 async def load_name(message: types.Message, state: FSMContext):
     await state.update_data(name=message.text)
     data = await state.get_data()
     try:
-        open("parser/login/users.json", "x").close()
+        open(settings.users_file_path, "x").close()
     except FileExistsError:
         pass
-    with open("parser/login/users.json", "r") as f:
+    with open(settings.users_file_path, "r") as f:
         try:
             users = json.load(f)
         except json.decoder.JSONDecodeError:
             users = dict()
-        users.update({data.get('name'): {"login": data.get('login'), "password": data.get('password'), "cookies": {"idp_srv_id": "", "srv_id": "", "JSESSIONID": "", "shib_idp_session": ""}}})
+        users.update({data.get('name'): {"login": data.get('login'), "password": data.get('password'), "cookies": {}}})
     with open("parser/login/users.json", "w") as f:
         json.dump(users, f, indent=4)
         await message.reply(f"Пользователь {data.get('name')} добавлен", reply_markup=buttons.main_menu)
     await state.finish()
 
 
-async def start_delete_user(message: types.Message):
-    await DeleteUserState.name.set()
-    await message.reply('Введите имя')
+async def delete_user_start(message: types.Message):
+    await DeleteUserState.user_name.set()
+    await message.reply('Выберите пользователя', reply_markup=buttons.users_menu)
 
 
-async def delete_user(message: types.Message, state: FSMContext):
+async def delete_user_confirm(message: types.Message, state: FSMContext):
     await state.update_data(name=message.text)
-    with open("users.json", "r") as f:
-        users = json.load(f)
+    await DeleteUserState.confirm.set()
+    await message.reply(f"Точно удалить пользователя {message.text}?", reply_markup=buttons.confirm_menu)
+
+
+async def delete_user_finish(message: types.Message, state: FSMContext):
+    await state.update_data(confirm=message.text)
+    data = await state.get_data()
+    if data.get("confirm").lower() == "да" or data.get("confirm").lower() == "yes":
         try:
-            users.pop(message.text)
-        except KeyError:
-            await message.reply("Не верное имя")
+            with open(settings.users_file_path, "r") as f:
+                users = json.load(f)
+                try:
+                    users.pop(data.get("name"))
+                except KeyError:
+                    await message.reply("Имя отсутствует в списке")
+                    return
+            with open(settings.users_file_path, "w") as f:
+                json.dump(users, f, indent=4)
+        except json.decoder.JSONDecodeError:
+            await message.answer("Нет зарегистрированных пользователей", reply_markup=buttons.main_menu)
+            await state.finish()
             return
-    with open("users.json", "w") as f:
-        json.dump(users, f, indent=4)
-    await message.reply(f"{message.text} удален(а)", reply_markup=buttons.main_menu)
-    await state.finish()
-
-
-# async def create_schedule_task(message: types.Message):
-#     user_tasks = scheduler.get_user_tasks(message.from_user.id)
-#     if user_tasks:
-#         logger.info('Попытка запуска периодических тасок. Уже запущено')
-#         await message.answer(text="Уже запущено")
-#     else:
-#         await scheduler.job_create(function=get_requests, params=[message, True])
-#         logger.info('Периодические таски запущены')
-#         await message.answer(text="Запустилось")
+        logger.info(f"{data.get('name')} удален(а)")
+        await message.reply(f"{data.get('name')} удален(а)", reply_markup=buttons.main_menu)
+        await state.finish()
+    elif data.get("confirm").lower() == "нет" or data.get("confirm").lower() == "no":
+        await state.finish()
+        await message.answer("Отмена", reply_markup=buttons.main_menu)
+    else:
+        await message.reply("Выберите: да(yes) или нет(no)", reply_markup=buttons.confirm_menu)
 
         
-async def cancel_schedule_task(message: types.Message):
+async def cancel_periodic_parse(message: types.Message):
     user_tasks = scheduler.get_user_tasks(message.from_user.id)
     if len(user_tasks) > 0:
         scheduler.job_remove(user_tasks)
@@ -226,13 +248,21 @@ async def start_response(message: types.Message):
 
 async def help_response(message: types.Message):
     help_message = "Парсер наличия заявок для ГВЭ в Меркурии.\n" + \
-        "1. \"Получить список активных заявок\" - возвращает список ферм с новыми заявками транзакций; \n" + \
+        "1. \"Получить список активных заявок\" - возвращает список ферм с заявками транзакций; \n" + \
         "2. \"Запустить периодически\" - п.1, проверка раз в несколько минут; \n" + \
         "3. \"Остановить\" - остановка периодического выполнения \n" \
         "/adduser - добавить пользователя \n" \
-        "/deleteuser - удалить пользователя \n"
-    await message.reply(text=help_message,
-                        reply_markup=buttons.main_menu,
-                        )
+        "/deleteuser - удалить пользователя \n" \
+        "/cancel - сброс состояния \n"
+    await message.reply(text=help_message, reply_markup=buttons.main_menu)
+
+
+async def cancel_handler(message: types.Message, state: FSMContext):
+    current_state = await state.get_state()
+    logger.info('Сброс состояния')
+    if current_state is None:
+        return
+    await state.finish()
+    await message.reply("Ок~", reply_markup=buttons.main_menu)
